Make useConfirmDirtyLeave its own hook
